Finalizacao do projeto 19/04 18:20

+++ Domain/Product.cs
@@ -7,8 +7,8 @@ namespace Domain
         public string Name { get; set; }
         public string Description { get; set; }
         public DateTime CreatedAt { get; set; }
-        public decimal Price { get; set; }
-        public decimal Quantity { get; set; }
+        public double Price { get; set; }
+        public double Quantity { get; set; }
 
         public Product()
         {
